pyproject.toml
[tool.poetry]
name = "repl_python3_beta_cerebrus"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
async-timeout = "^3.0.1"
bs4 = "^0.0.1"
flask = "^2.0.1"
opencv-python = "^4.5.2"
pytz = "^2021.1"
anime-downloader = "^5.0.9"
replit = "^3.1.0"
requests-oauthlib = "^1.3.0"
wikipedia = "^1.4.0"
scikit-image = "^0.18.1"
youtube-dl = "^2021.5.16"
pydealer = "^1.4.0"
dhooks = "^1.1.4"
translators = "^4.7.16"
aiohttp = "^3.7.4"
spotipy = "^2.18.0"
discord = "^1.0.1"
python-dotenv = "^0.17.1"
asyncpg = "^0.23.0"
TenGiphPy = "^2.5.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus.git"}
pyjokes = "^0.6.0"
names = "^0.3.0"
"akinator.py" = "^5.0.0"
prsaw = "^0.3.1"
jishaku = "^1.20.0"
discord-ext-slash = "^0.7.0"
discord-py-slash-command = "^1.1.2"
colorthief = "^0.2.1"
timeago = "^1.0.15"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
